# backend/src/main/kotlin/com/leaguelift/authorization/web/AuthorizationDto.kt
package com.leaguelift.authorization.web

import com.leaguelift.authorization.domain.AuthorizationContext
import java.util.UUID

data class ContextResponse(
	val contextType: String,
	val resourceId: UUID?,
	val organizationId: UUID?,
	val label: String,
	val role: String,
	val capabilities: List<String>,
)

fun AuthorizationContext.toResponse() = ContextResponse(
	contextType = contextType.name,
	resourceId = resourceId,
	organizationId = organizationId,
	label = label,
	role = role,
	capabilities = capabilities.sorted(),
)

data class GrantRoleAssignmentRequest(
	val userId: UUID,
	val role: String,
)

data class RoleAssignmentResponse(
	val id: UUID,
	val userId: UUID,
	val userEmail: String?,
	val userDisplayName: String?,
	val contextType: String,
	val resourceId: UUID?,
	val role: String,
)