# docs/openapi.yaml
openapi: 3.0.3
info:
  title: LeagueLift API
  description: >
    LeagueLift revenue, fundraising, commerce, and fee-management API for youth sports
    organizations. This document covers Phase 0 (foundation), Phase 1 (organization
    onboarding, teams, tournaments, and public pages), Phase 2 (households,
    participants, fee templates, and fee assignments), Phase 3 slice 11
    (fundraising campaigns and their public pages) plus its remainder (campaign
    contribution recording via test-mode Stripe Checkout, real raised totals, and
    an inbound Stripe webhook receiver — DESIGN-DOC.md section 8.5; family
    credit rules remain a later slice pending section 19.3's open questions),
    traditional email/password authentication (superseding
    the originally planned external-IdP/Auth0 approach — see DESIGN-DOC.md section
    11.5), Phase 1's remaining file-upload/branding slice (organization logo/cover
    image via S3-compatible object storage — DESIGN-DOC.md section 11, ADR-012), and
    Phase 2's remainder (manual/offline fee payments, manual discounts/credits,
    org-wide collections listing and CSV export, and Stripe Connect Express
    onboarding-only scaffolding — DESIGN-DOC.md section 14.3, ADR-005 — no live
    charge routing yet), and Phase 4 slice 1 (apparel commerce proof: stores,
    Printify-catalog-backed products/variants with real Printify cost snapshots,
    a Printify catalog read proxy, multi-item test-mode Stripe Checkout with US
...
servers:
...
paths:
...
  /me/contexts:
    get:
      tags: [authorization]
      summary: Every authorization context (PERSONAL/ATHLETE/HOUSEHOLD/TEAM/ORGANIZATION/TOURNAMENT/PLATFORM_ADMIN) the caller holds, with resolved capabilities (Phase 7/ADR-020)
      operationId: getMyContexts
      responses:
        "200":
          description: Contexts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ContextResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /organizations/{organizationId}/teams/{teamId}/role-assignments:
    get:
      tags: [authorization]
      summary: List active role assignments for a team (org OWNER/ADMINISTRATOR only; backs the "Manage access" admin UI)
      operationId: listTeamRoleAssignments
      parameters:
        - $ref: "#/components/parameters/OrganizationIdParam"
        - $ref: "#/components/parameters/TeamIdParam"
      responses:
        "200":
          description: Active role assignments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RoleAssignmentResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    post:
      tags: [authorization]
      summary: Grant a team-scoped role (COACH_READ/TEAM_EDITOR/TEAM_MANAGER) to an org member (org OWNER/ADMINISTRATOR only; Phase 7/ADR-020)
      operationId: grantTeamRoleAssignment
      parameters:
        - $ref: "#/components/parameters/OrganizationIdParam"
        - $ref: "#/components/parameters/TeamIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GrantRoleAssignmentRequest"
      responses:
        "201":
          description: Role assignment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoleAssignmentResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /organizations/{organizationId}/teams/{teamId}/role-assignments/{assignmentId}:
    delete:
      tags: [authorization]
      summary: Revoke a team-scoped role assignment (org OWNER/ADMINISTRATOR only; Phase 7/ADR-020)
      operationId: revokeTeamRoleAssignment
      parameters:
        - $ref: "#/components/parameters/OrganizationIdParam"
        - $ref: "#/components/parameters/TeamIdParam"
        - name: assignmentId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Revoked
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /organizations/{organizationId}/tournaments/{tournamentId}/role-assignments:
    get:
      tags: [authorization]
      summary: List active role assignments for a tournament (org OWNER/ADMINISTRATOR only; backs the "Manage access" admin UI)
      operationId: listTournamentRoleAssignments
      parameters:
        - $ref: "#/components/parameters/OrganizationIdParam"
        - $ref: "#/components/parameters/TournamentIdParam"
      responses:
        "200":
          description: Active role assignments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RoleAssignmentResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    post:
      tags: [authorization]
      summary: Grant a tournament-scoped role (TOURNAMENT_VIEWER/TOURNAMENT_ADMINISTRATOR) to an org member (org OWNER/ADMINISTRATOR only; Phase 7/ADR-020)
      operationId: grantTournamentRoleAssignment
      parameters:
        - $ref: "#/components/parameters/OrganizationIdParam"
        - $ref: "#/components/parameters/TournamentIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GrantRoleAssignmentRequest"
      responses:
        "201":
          description: Role assignment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoleAssignmentResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /organizations/{organizationId}/tournaments/{tournamentId}/role-assignments/{assignmentId}:
    delete:
      tags: [authorization]
      summary: Revoke a tournament-scoped role assignment (org OWNER/ADMINISTRATOR only; Phase 7/ADR-020)
      operationId: revokeTournamentRoleAssignment
      parameters:
        - $ref: "#/components/parameters/OrganizationIdParam"
        - $ref: "#/components/parameters/TournamentIdParam"
        - name: assignmentId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Revoked
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    OrganizationIdParam:
      name: organizationId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    TeamIdParam:
      name: teamId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    TournamentIdParam:
      name: tournamentId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: Validation failure
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Missing or invalid credentials
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Authenticated but not authorized for this resource
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Resource not found (or not visible to the caller)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
...
    ErrorResponse:
      type: object
      required: [code, message, requestId, fieldErrors]
      properties:
        code:
          type: string
          example: ORGANIZATION_NOT_FOUND
        message:
          type: string
          example: The organization could not be found.
        requestId:
          type: string
          example: req_01hz8x9y2z
        fieldErrors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
...
    ContextResponse:
      type: object
      description: One authorization context the caller holds (Phase 7/ADR-020, DESIGN-DOC.md section 4.2).
      required: [contextType, label, role, capabilities]
      properties:
        contextType:
          type: string
          enum: [PERSONAL, ATHLETE, HOUSEHOLD, TEAM, ORGANIZATION, TOURNAMENT, PLATFORM_ADMIN]
        resourceId:
          type: string
          format: uuid
          nullable: true
        organizationId:
          type: string
          format: uuid
          nullable: true
        label:
          type: string
        role:
          type: string
        capabilities:
          type: array
          items:
            type: string
          description: Resolved, deny-by-default capability strings for this context — never re-derive from `role` client-side.
...
    GrantRoleAssignmentRequest:
      type: object
      required: [userId, role]
      properties:
        userId:
          type: string
          format: uuid
        role:
          type: string
          description: One of COACH_READ/TEAM_EDITOR/TEAM_MANAGER (team endpoint) or TOURNAMENT_VIEWER/TOURNAMENT_ADMINISTRATOR (tournament endpoint).
    RoleAssignmentResponse:
      type: object
      required: [id, userId, contextType, role]
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        userEmail:
          type: string
          nullable: true
        userDisplayName:
          type: string
          nullable: true
        contextType:
          type: string
        resourceId:
          type: string
          format: uuid
          nullable: true
        role:
          type: string
...
security:
  - bearerAuth: []